+++ extensions/tekton/mock/src/main/java/io/fabric8/tekton/mock/TektonServer.java
@@ -17,26 +17,23 @@ package io.fabric8.tekton.mock;
 
 import io.fabric8.kubernetes.client.server.mock.KubernetesCrudDispatcher;
 import io.fabric8.mockwebserver.Context;
-import io.fabric8.mockwebserver.ServerRequest;
-import io.fabric8.mockwebserver.ServerResponse;
 import io.fabric8.mockwebserver.dsl.MockServerExpectation;
 import io.fabric8.tekton.client.TektonClient;
 import okhttp3.mockwebserver.MockWebServer;
 import org.junit.rules.ExternalResource;
 
 import java.util.HashMap;
-import java.util.Queue;
 
 public class TektonServer extends ExternalResource {
 
   protected TektonMockServer mock;
   private TektonClient client;
 
-  private boolean https;
+  private final boolean https;
   // In this mode the mock web server will store, read, update and delete
   // kubernetes resources using an in memory map and will appear as a real api
   // server.
-  private boolean crudMode;
+  private final boolean crudMode;
 
   public TektonServer() {
     this(true, false);
@@ -51,6 +48,7 @@ public class TektonServer extends ExternalResource {
     this.crudMode = crudMode;
   }
 
+  @Override
   public void before() {
     mock = crudMode
       ? new TektonMockServer(new Context(), new MockWebServer(), new HashMap<>(), new KubernetesCrudDispatcher(), true)
@@ -59,6 +57,7 @@ public class TektonServer extends ExternalResource {
     client = mock.createTekton();
   }
 
+  @Override
   public void after() {
     mock.destroy();
     client.close();
@@ -68,7 +67,6 @@ public class TektonServer extends ExternalResource {
   public TektonClient getTektonClient() {
     return client;
   }
-
 
   public MockServerExpectation expect() {
     return mock.expect();
@@ -82,9 +80,5 @@ public class TektonServer extends ExternalResource {
   @Deprecated
   public void expectAndReturnAsString(String path, int code, String body) {
     expect().withPath(path).andReturn(code, body).always();
-  }
-
-  public MockWebServer getMockServer() {
-    return mock.getServer();
   }
 }